requirements.txt
asgiref==3.2.3
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
dateparser==0.7.2
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.3
django-import-export==2.0.2
django-modeltranslation==0.14.4
django-tinymce==2.8.0
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
et-xmlfile==1.0.1
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-speech==1.3.2
googleapis-common-protos==1.51.0
grpcio==1.33.2
idna==2.9
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2019.3
regex==2020.2.20
requests==2.23.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
tablib==1.1.0
tzlocal==2.0.0
urllib3==1.25.8
